backend/requirements.txt
# --- Core FastAPI Framework ---
fastapi==0.115.2
uvicorn[standard]==0.32.0

# --- Database & ORM ---
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9        # PostgreSQL driver
python-dotenv==1.0.1

# --- Pydantic & Utilities ---
pydantic==2.9.2
pydantic-settings==2.5.2

# --- Date & Time handling ---
python-dateutil==2.9.0.post0
pytz==2024.1

# --- Optional / Useful for logging & production ---
loguru==0.7.2

# --- Testing / Development (optional) ---
httpx==0.27.2
pytest==8.3.3
